sensors: update Multihal For N
Remove header from LOCAL_SRC_FILES

The build system does not use headers in LOCAL_SRC_FILES. To prevent
typos, it's now a warning to add something to LOCAL_SRC_FILES that
is not used.

Change-Id: Ieb6c6d383c04005077043d04f1b4985964ef0341

Sensors: MultiHal: Increase xfer buffer size

Increased the size of the sensor event transfer queue from 20 to
36 sensor events.  This size now matches Sensor Services buffer size
that it passes to poll() for reading events and will allow the
multihal to transfer the maximum events possible per poll cycle.

Change-Id: I55459ea0a70b9da848d3d3d43590050d71c53e28